==> scripts/00_microbiome_analysis.R <==
suppressMessages({
  suppressWarnings({
    source(file = "scripts/01_install_load_packages.R")
    source(file = "scripts/02_import_clean_data.R")
    source(file = "scripts/03_a_rarefaction_curves.R")
    source(file = "scripts/03_b_feature_accumulation_curve.R")
    #source(file = "scripts/04_quality_control.R")
    source(file = "scripts/05_species_function_curve.R")
    source(file = "scripts/06_prevalence.R")
    source(file = "scripts/07_prevalence_tree.R")
    source(file = "scripts/08_barplot_community_composition_bacteria.R")
    #source(file = "scripts/09_abundance_species_bacteria.R")
    source(file = "scripts/10_function_composition.R")
    source(file = "scripts/11_cross_association_phyla_func_cat.R")
    source(file = "scripts/12_feature_importance_analysis.R")
    source(file = "scripts/13_combine_plots_unknown_accumulation.R")
    source(file = "scripts/14_combine_plots_prevalence_RE.R")
    source(file = "scripts/15_combine_plots_heataps.R")
  
  })
})
